+++ lib/src/circle_ci/models/project_yaml_config_model.dart
@@ -1,32 +1,27 @@
-import 'package:flutter_monorepo_build_tools/src/circle_ci/models/orbs_model.dart';
-import 'package:flutter_monorepo_build_tools/src/circle_ci/models/worflows_model.dart';
+import 'package:flutter_monorepo_build_tools/src/circle_ci/models/workflow_model.dart';
 
 class ProjectYamlConfigModel {
-  double? version;
-  bool? setup;
-  OrbsModel? orbs;
-  WorkflowsModel? workflows;
+  Map<String, dynamic>? params;
+  List<WorkflowModel>? workflows;
 
-  ProjectYamlConfigModel({this.version, this.setup, this.orbs, this.workflows});
+  ProjectYamlConfigModel({this.params, this.workflows});
 
   ProjectYamlConfigModel.fromJson(Map<String, dynamic> json) {
-    version = json['version'];
-    setup = json['setup'];
-    orbs = json['orbs'] != null ? OrbsModel.fromJson(json['orbs']) : null;
-    workflows = json['workflows'] != null
-        ? WorkflowsModel.fromJson(json['workflows'])
+    workflows = (json['workflows'] != null)
+        ? json['workflows']
+            .keys
+            .map<WorkflowModel>(
+                (k) => WorkflowModel.fromJson(json['workflows'][k])..name = k)
+            .toList()
         : null;
+    json.remove('workflows');
+    params = json;
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['version'] = version;
-    data['setup'] = setup;
-    if (orbs != null) {
-      data['orbs'] = orbs!.toJson();
-    }
+    final Map<String, dynamic> data = params ?? {};
     if (workflows != null) {
-      data['workflows'] = workflows!.toJson();
+      data['workflows'] = {for (var w in workflows!) w.name: w.toJson()};
     }
     return data;
   }